Configure Strip URI parameter for a set of API Objects via Ingress Annotations
The [`strip_uri`](https://getkong.org/docs/0.10.x/proxy/#the-strip_uri-property) property determines whether or not to strip a
matching prefix from the upstream URI to be requested. The default value is `true`
but it might not be the desire behavior in all cases.

This PR enables an Ingress Resource to be annotated as

```
...
annotations:
  ingress.kubernetes.io/strip-uri: "[false | true]"
```

The controller then reads this annotation and uses its value while creating
the different API objects via Kong Admin.